test(ui): move the TreeNode tests into TreeNode.test.tsx
They were appended to PoolTree.test.tsx, which covers a different component in
a different directory, so nobody editing wizard/components/TreeNode.tsx would
find them by filename. Pure move - same three assertions.
